+++ packages/types/custom/core.ts
@@ -19,7 +19,7 @@ import {
   SubscribeOptions,
 } from 'vuex'
 import { Feature as GeoJsonFeature, FeatureCollection } from 'geojson'
-import Vue, { VueConstructor, WatchOptions } from 'vue'
+import { VueConstructor, WatchOptions } from 'vue'
 
 /**
  *
@@ -34,8 +34,22 @@ export interface PluginOptions {
 
 export type RenderType = 'iconMenu' | 'independent' | 'footer'
 
+export type LoaderStyles =
+  | 'CircleLoader'
+  | 'BasicLoader'
+  | 'none'
+  | 'RingLoader'
+  | 'RollerLoader'
+  | 'SpinnerLoader'
+  | 'v-progress-linear'
+
+/** LoadingIndicator Module Configuration */
+export interface LoadingIndicatorConfiguration extends PluginOptions {
+  loaderStyle?: LoaderStyles
+}
+
 /** Possible search methods by type */
-export type SearchType = 'bkg' | 'gazetteer' | 'wfs' | 'mpapi' | string
+export type SearchType = 'bkg' | 'wfs' | 'mpapi' | string
 
 /**
  * Additional queryParameters for the GET-Request;
@@ -69,13 +83,13 @@ export type SearchMethodFunction = (
 ) => Promise<FeatureCollection> | never
 
 export interface SelectResultPayload {
-  feature: GeoJsonFeature
+  feature: GeoJsonFeature & { title: string }
   categoryId: number
 }
 
-export type SelectResultFunction = (
-  PolarActionContext,
-  SelectResultPayload
+export type SelectResultFunction<S, G> = (
+  context: PolarActionContext<S, G>,
+  payload: SelectResultPayload
 ) => void
 
 /**
@@ -107,8 +121,10 @@ export interface AddressSearchConfiguration extends PluginOptions {
   categoryProperties?: Record<string, AddressSearchCategoryProperties>
   // optional additional search methods (client-side injections)
   customSearchMethods?: Record<string, SearchMethodFunction>
+  /** NOTE regarding \<any, any\> – skipping further type chain upwards precision due to object optionality/clutter that would continue to MapConfig level; the inverted rabbit hole ends here; not using "unknown" since that errors in client configuration, not using "never" since that errors in AddressSearch plugin; this way, "any"thing goes */
   // optional selectResult overrides (client-side injections)
-  customSelectResult?: Record<string, SelectResultFunction>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  customSelectResult?: Record<string, SelectResultFunction<any, any>>
   focusAfterSearch?: boolean
   // definition of groups referred to in searchMethods
   groupProperties?: Record<string, AddressSearchGroupProperties>
@@ -167,6 +183,7 @@ export interface DrawConfiguration extends Partial<PluginOptions> {
   selectableDrawModes?: DrawMode[]
   style?: DrawStyle
   textStyle?: TextStyle
+  enableOptions?: boolean
 }
 
 export interface ExportConfiguration extends PluginOptions {
@@ -341,7 +358,7 @@ export interface GfiConfiguration extends PluginOptions {
    * Optionally replace GfiContent component.
    * Usable to completely redesign content of GFI window.
    */
-  gfiContentComponent?: Vue
+  gfiContentComponent?: VueConstructor
   /**
    * Limits the viewable GFIs per layer by this number. The first n elements
    * are chosen arbitrarily. Useful if you e.g. just want one result, or to
@@ -587,6 +604,7 @@ export interface MapConfig {
   startResolution?: number
   vuetify?: UserVuetifyPreset
   addressSearch?: AddressSearchConfiguration
+  loadingIndicator?: LoadingIndicatorConfiguration
   attributions?: AttributionsConfiguration
   draw?: DrawConfiguration
   export?: ExportConfiguration
@@ -621,7 +639,7 @@ type MoveHandleProps = object
 export interface MoveHandleProperties {
   closeLabel: string
   closeFunction: (...args: unknown[]) => unknown
-  component: Vue
+  component: VueConstructor
   // Plugin that added the moveHandle
   plugin: string
   closeIcon?: string
@@ -629,7 +647,7 @@ export interface MoveHandleProperties {
 }
 
 export interface MoveHandleActionButton {
-  component: Vue
+  component: VueConstructor
   props?: MoveHandleProps
 }
 
@@ -647,7 +665,9 @@ export interface CoreState {
   mapHasDimensions: boolean
   moveHandle: number
   moveHandleActionButton: number
-  plugin: object
+  // NOTE truly any since external plugins may bring whatever; unknown will lead to further errors
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  plugin: Record<string, any>
   selected: number
   zoomLevel: number
 }